requirements.txt
asgiref==3.4.1
bandit==1.7.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
djangorestframework==3.12.4
flake8==4.0.1
gitdb==4.0.9
GitPython==3.1.24
idna==3.3
importlib-metadata==4.2.0
install==1.3.4
mccabe==0.6.1
pbr==5.7.0
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
smmap==5.0.0
sqlparse==0.4.2
stevedore==3.5.0
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
zipp==3.6.0
